Transaction Post and Gets Now Sends through Items

+++ backend/src/Models/TransItem.cs
@@ -7,7 +7,7 @@ namespace Pz.Cheeseria.Api.Models
         public int ItemNo { get; set; } // Identity
         public int CheeseId { get; set; } // FK to Product Sold
         public int Quantity { get; set; }
-        public int Price { get; set; }
-        public int Total { get; set; }
+        public decimal Price { get; set; }
+        public decimal Total { get; set; }
     }
 }
